Move GaslessVotingToggle to GaslessVoting folder
